Seemed to have gotten the riff container set up, but now I need to see if it'll actually be imported into a decoder.

+++ src/audio.rs
@@ -72,7 +72,7 @@ pub fn save(audio: &AudioBuffer, path: &Path) -> AudioResult<bool> {
 
 /// Trait which all decoders must implement in order to return an `AudioBuffer` and metadata
 pub trait AudioDecoder {
-  fn bit_rate(&self) -> AudioResult<u8>;
+  fn bit_rate(&self) -> AudioResult<u32>;
   fn sample_rate(&self) -> AudioResult<u32>;
   fn channels(&self) -> AudioResult<u32>;
   fn sample_order(&self) -> AudioResult<SampleOrder>;
