+++ src/buffer.rs
@@ -9,7 +9,7 @@ pub enum SampleOrder { MONO, INTERLEAVED, REVERSED, PLANAR }
 /// Holds all samples and necessary audio data for processing and encoding.
 #[derive(Clone, Debug)]
 pub struct AudioBuffer {
-  pub bit_rate: u8,
+  pub bit_rate: u32,
   pub sample_rate: u32,
   pub channels: u32,
   pub order: SampleOrder,
